3.0

+++ qr-menu-underground/database/migrations/2024_08_02_181202_create_carts_table.php
@@ -17,10 +17,9 @@ return new class extends Migration
             $table->unsignedInteger('table_number');
             $table->unsignedBigInteger('product_id');
             $table->integer('quantity');
-            $table->decimal('price');
+            $table->integer('price');
             $table->string('device_info')->nullable();
             $table->timestamps();
-
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
         });
     }
